Create IDW
Inverse Distance Weighting (IDW) is a method for interpolating values from scattered data points by giving more weight to data points that are closer to the interpolation point. The basic idea behind IDW is that the value at an interpolation point should be a weighted average of the values at nearby data points, where the weight of each data point is inversely proportional to its distance from the interpolation point.